match: AICopManager::UpdatePursuits in main/Speed/Indep/SourceLists/zAI
- Cast incremented mTotalCopsDestroyed to float for GManager::TrackValue overload
- Achieves 100.0% matching assembly (2516 bytes)
- Zero regressions across all source units

+++ src/Speed/Indep/Src/AI/Activities/AICopManager.cpp
@@ -1854,7 +1854,7 @@ void AICopManager::UpdatePursuits() {
             }
 
             if (IsDestroyed && vehicleRemoved && bAllowStatsToAccumulate) {
-                GManager::Get().TrackValue("cops_destroyed", ++this->mTotalCopsDestroyed);
+                GManager::Get().TrackValue("cops_destroyed", (float)++this->mTotalCopsDestroyed);
             }
         } else {
             if (ivehicle->GetVehicleClass() == VehicleClass::CHOPPER) {
